implement mobile designs and match ds breakpoints
