+++ src/ui/nypl-base-theme/foundations/breakpoints.ts
@@ -1,9 +1,9 @@
 // theme.js
 const breakpoints = {
-  sm: '20em', // 320px
-  md: '38em', // 600px
-  lg: '60em', // 960px
-  xl: '80em', // 1280px
+  sm: '30em',
+  md: '48em',
+  lg: '64em',
+  xl: '80em',
   '2xl': '96em',
 };
